fix parseAbstractPos...

+++ src/meowtro/metro_system/railway/RailwayRealizer.java
@@ -4,5 +4,5 @@ import meowtro.Position;
 
 public interface RailwayRealizer {
     public int parsePositionToAbstractPosition(Position p); 
-    public Position parseAbstractPositionToPosition(int abstractPosition); 
+    public Position parseAbstractPositionToPosition(double abstractPosition); 
 }
